added ability to update existing pr comment with report

+++ src/index.js
@@ -6,6 +6,7 @@ import { parse } from "./lcov"
 import { diff } from "./comment"
 import { getChangedFiles } from "./get_changes"
 import { deleteOldComments } from "./delete_old_comments"
+import { updateOldComment } from "./update_old_comment"
 import { normalisePath } from "./util"
 
 const MAX_COMMENT_CHARS = 65536
@@ -19,6 +20,8 @@ async function main() {
 		core.getInput("filter-changed-files").toLowerCase() === "true"
 	const shouldDeleteOldComments =
 		core.getInput("delete-old-comments").toLowerCase() === "true"
+	const shouldUpdateComment =
+		core.getInput("update-comment").toLowerCase() === "true"
 	const title = core.getInput("title")
 
 	const raw = await fs.readFile(lcovFile, "utf-8").catch(err => null)
@@ -60,17 +63,27 @@ async function main() {
 	const baselcov = baseRaw && (await parse(baseRaw))
 	const body = diff(lcov, baselcov, options).substring(0, MAX_COMMENT_CHARS)
 
-	if (shouldDeleteOldComments) {
+	if (shouldDeleteOldComments && !shouldUpdateComment) {
 		await deleteOldComments(githubClient, options, context)
 	}
 
 	if (context.eventName === "pull_request") {
-		await githubClient.issues.createComment({
-			repo: context.repo.repo,
-			owner: context.repo.owner,
-			issue_number: context.payload.pull_request.number,
-			body: body,
-		})
+
+		const hasUpdatedComment = false;
+
+		if (shouldUpdateComment) {
+			hasUpdatedComment = await updateOldComment(githubClient, options, context)
+		}
+
+		if (!hasUpdatedComment) {
+			await githubClient.issues.createComment({
+				repo: context.repo.repo,
+				owner: context.repo.owner,
+				issue_number: context.payload.pull_request.number,
+				body: body,
+			})
+		}
+
 	} else if (context.eventName === "push") {
 		await githubClient.repos.createCommitComment({
 			repo: context.repo.repo,
